requirements.txt
# Required dependencies
python-dotenv>=1.0.0
langchain>=0.0.335
langchain-core>=0.1.9
langchain-openai>=0.0.2
openai>=1.3.0
faiss-cpu>=1.7.4
pdfplumber>=0.10.2
docx2txt>=0.8
python-magic>=0.4.27
rich>=13.5.2
tqdm>=4.66.1
# Additional tools
requests>=2.31.0
typing-extensions>=4.8.0
